+++ starport/cmd/account_list.go
@@ -1,0 +1,34 @@
+package starportcmd
+
+import (
+	"github.com/spf13/cobra"
+	"github.com/tendermint/starport/starport/pkg/cosmosaccount"
+)
+
+func NewAccountList() *cobra.Command {
+	c := &cobra.Command{
+		Use:   "list",
+		Short: "List accounts",
+		RunE:  accountListHandler,
+	}
+
+	c.Flags().AddFlagSet(flagSetKeyringBackend())
+	c.Flags().AddFlagSet(flagSetAccountPrefixes())
+
+	return c
+}
+
+func accountListHandler(cmd *cobra.Command, args []string) error {
+	ca, err := cosmosaccount.New(getKeyringBackend(cmd))
+	if err != nil {
+		return err
+	}
+
+	accounts, err := ca.List()
+	if err != nil {
+		return err
+	}
+
+	printAccounts(cmd, accounts...)
+	return nil
+}