+++ starport/cmd/cmd.go
@@ -59,6 +59,7 @@ starport scaffold chain github.com/cosmonaut/mars`,
 	c.AddCommand(NewChain())
 	c.AddCommand(NewGenerate())
 	c.AddCommand(NewNetwork())
+	c.AddCommand(NewAccount())
 	c.AddCommand(NewRelayer())
 	c.AddCommand(NewTools())
 	c.AddCommand(NewDocs())