+++ starport/cmd/scaffold_message.go
@@ -9,6 +9,8 @@ import (
 	"github.com/tendermint/starport/starport/services/scaffolder"
 )
 
+const flagSigner = "signer"
+
 // NewScaffoldMessage returns the command to scaffold messages
 func NewScaffoldMessage() *cobra.Command {
 	c := &cobra.Command{
@@ -21,6 +23,7 @@ func NewScaffoldMessage() *cobra.Command {
 	c.Flags().String(flagModule, "", "Module to add the message into. Default: app's main module")
 	c.Flags().StringSliceP(flagResponse, "r", []string{}, "Response fields")
 	c.Flags().StringP(flagDescription, "d", "", "Description of the command")
+	c.Flags().String(flagSigner, "", "Label for the message signer (default: creator)")
 
 	return c
 }
@@ -29,33 +32,30 @@ func messageHandler(cmd *cobra.Command, args []string) error {
 	s := clispinner.New().SetText("Scaffolding...")
 	defer s.Stop()
 
-	// Get the module to add the type into
-	module, err := cmd.Flags().GetString(flagModule)
-	if err != nil {
-		return err
-	}
+	var (
+		module, _    = cmd.Flags().GetString(flagModule)
+		resFields, _ = cmd.Flags().GetStringSlice(flagResponse)
+		desc, _      = cmd.Flags().GetString(flagDescription)
+		signer, _    = cmd.Flags().GetString(flagSigner)
+	)
 
-	// Get response fields
-	resFields, err := cmd.Flags().GetStringSlice(flagResponse)
-	if err != nil {
-		return err
-	}
+	var options []scaffolder.MessageOption
 
 	// Get description
-	desc, err := cmd.Flags().GetString(flagDescription)
-	if err != nil {
-		return err
+	if desc != "" {
+		options = append(options, scaffolder.WithDescription(desc))
 	}
-	if desc == "" {
-		// Use a default description
-		desc = fmt.Sprintf("Broadcast message %s", args[0])
+
+	// Get signer
+	if signer != "" {
+		options = append(options, scaffolder.WithSigner(signer))
 	}
 
 	sc, err := scaffolder.New(appPath)
 	if err != nil {
 		return err
 	}
-	sm, err := sc.AddMessage(placeholder.New(), module, args[0], desc, args[1:], resFields)
+	sm, err := sc.AddMessage(placeholder.New(), module, args[0], args[1:], resFields, options...)
 	if err != nil {
 		return err
 	}