+++ changelog.md
@@ -4,7 +4,14 @@
 
 ### Features
 
+- Added `--signer` flag to `scaffold message` to allow customising the name of the signer of the message
+
+## `v0.17.2`
+
+### Features
+
 - `client.toml` is initialized and used by node's CLI, can be configured through `config.yml` with the `init.client` property
+- Support serving Cosmos SDK `v0.43.x` based chains.
 
 ## `v0.17.1`